+++ src/coins.cpp
@@ -77,6 +77,9 @@ bool CCoinsViewBacked::BatchWrite(CCoinsMap &mapCoins,
 CCoinsViewCursor *CCoinsViewBacked::Cursor() const {
     return base->Cursor();
 }
+size_t CCoinsViewBacked::EstimateSize() const {
+    return base->EstimateSize();
+}
 
 SaltedTxidHasher::SaltedTxidHasher()
     : k0(GetRand(std::numeric_limits<uint64_t>::max())),